scripts/cli/requirements.txt
# CLI Dependencies
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0